+++ apps/docs/next.config.mjs
@@ -6,6 +6,7 @@ const withNextra = nextra({
     codeblocks: false,
   },
   defaultShowCopyCode: true,
+  contentDirBasePath: "/",
 });
 
 export default withNextra({